requirements.txt
asgiref==3.8.1
Django==5.1.3
psycopg[binary]==3.2.3
sqlparse==0.5.2
tzdata==2024.2
djangorestframework
httpx